refac: move search logic into `SearchResults` comp
What?
- Extract search logic and related comps from 'home' page into
`SearchResults` comp.
- Make `DhatuList` comp simpler.

Why?
To reduce clutter in 'home' page and 'DhatuList' comp.
